Improve support for floating labels
